# public/openapi.yaml
...
openapi: 3.0.3
info:
  title: AccessGate API
  version: 0.1.0
  description: Event management and ticketing system API
  contact:
    name: API Support
    email: azeezwaris17@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Development server
...
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: Get all events with filtering and pagination
      description: Retrieve all events for the authenticated admin with optional filtering and pagination
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, completed, cancelled]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Successfully retrieved events
          content:
            application/json:
              example:
                success: true
                events: []
                pagination:
                  currentPage: 1
                  totalPages: 1
                  totalEvents: 0
                  hasNext: false
                  hasPrev: false
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /events/create-event:
    post:
      tags:
        - Events
      summary: Create new event
      description: Create a new event with ticket types
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
            example:
              name: "Tech Conference 2024"
              date: "2024-06-15"
              time: "09:00"
              location: "Convention Center"
              description: "Annual technology conference"
              ticketTypes:
                - name: "General Admission"
                  price: 100
                  quantity: 200
                  description: "Standard access"
      responses:
        '200':
          description: Successfully created event
          content:
            application/json:
              example:
                success: true
                message: "Event created successfully"
                event:
                  _id: "507f1f77bcf86cd799439013"
                  name: "Tech Conference 2024"
                  date: "2024-06-15"
                  time: "09:00"
                  location: "Convention Center"
                  status: "active"
        '400':
          description: Missing required fields or invalid ticket types
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /events/{eventId}:
    get:
      tags:
        - Event Operations
      summary: Get event details
      description: Retrieve detailed information for a specific event
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved event details
          content:
            application/json:
              example:
                success: true
                event:
                  _id: "507f1f77bcf86cd799439013"
                  name: "Tech Conference 2024"
                  date: "2024-06-15"
                  time: "09:00"
                  location: "Convention Center"
                  status: "active"
                  totalRevenue: 1000.00
                  totalTickets: 200
                  totalSold: 10
                  totalAvailable: 190
        '401':
          description: Unauthorized
        '404':
          description: Event not found or access denied
        '500':
          description: Internal server error
...
  /events/{eventId}/edit-event:
    put:
      tags:
        - Event Operations
      summary: Update event
      description: Update event details
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Successfully updated event
          content:
            application/json:
              example:
                success: true
                message: "Event updated successfully"
                event:
                  _id: "507f1f77bcf86cd799439013"
                  name: "Updated Conference Name"
                  date: "2024-06-15"
                  time: "10:00"
                  location: "New Venue"
                  status: "active"
        '400':
          description: Cannot edit completed events
        '401':
          description: Unauthorized
        '404':
          description: Event not found or access denied
        '500':
          description: Internal server error
...
  /events/{eventId}/get-selected-event-details:
    get:
      tags:
        - Event Operations
      summary: Get detailed event statistics
      description: Retrieve comprehensive statistics and details for a specific event
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved event details
          content:
            application/json:
              example:
                _id: "507f1f77bcf86cd799439013"
                name: "Tech Conference 2024"
                date: "2024-06-15"
                time: "09:00"
                location: "Convention Center"
                totalRevenue: 1000.00
                totalTickets: 200
                totalSold: 10
                totalAvailable: 190
                checkInCount: 5
                totalTicketCount: 10
                ticketTypes: []
        '401':
          description: Unauthorized
        '404':
          description: Event not found or access denied
        '500':
          description: Internal server error

  # ==================== ANALYTICS ENDPOINTS ====================
...
  /events/get-events-stat:
    get:
      tags:
        - Analytics
      summary: Get events statistics
      description: Retrieve statistics for all events
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved statistics
          content:
            application/json:
              example:
                success: true
                statistics:
                  totalEvents: 5
                  activeEvents: 3
                  completedEvents: 2
                  cancelledEvents: 0
                  totalRevenue: 2500.00
                  totalTicketsSold: 25
                  totalTicketsAvailable: 175
                  recentEvents: []
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # ==================== TICKET ENDPOINTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateEventRequest:
      type: object
      required:
        - name
        - date
        - time
        - location
        - description
        - ticketTypes
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: string
        description:
          type: string
        ticketTypes:
          type: array
          items:
            type: object
            required:
              - name
              - price
              - quantity
            properties:
              name:
                type: string
              price:
                type: number
                minimum: 0
              quantity:
                type: integer
                minimum: 1
              description:
                type: string
              specialConditions:
                type: string
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: string
        description:
          type: string
        ticketTypes:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
              quantity:
                type: integer
              description:
                type: string
              specialConditions:
                type: string
...